Move Button outside of unnecessary folder
