+++ src/components/pages/GenerationPage/GeneratePetitionModal/GeneratePetitionModal.js
@@ -10,7 +10,7 @@ import useKeyPress from '../../../../hooks/useKeyPress';
 
 // Children/Components
 import AgencyAutocomplete from '../GenerationInput/AgencyAutocomplete';
-import { Button, CloseButton } from '../../../elements/Button/Button.styled';
+import { Button, CloseButton } from '../../../elements/Button';
 import Axios from '../../../../service/axios';
 
 const ModalCloseButton = styled(CloseButton)`
